Partial cherry-pick: Refactor the link to interval to extract the generation of the uri and use URI to generate the urls and keep the parameters
This commit is a partial cherry-pick of b2690604. The original commit
does more than the brief description says. This commit does exactly
what the brief description says (although I should point that a change
in behavior, no matter how small, is not a refactoring).

Conflicts:

	lib/gdash/sinatra_app.rb
	views/layout.erb

The original commit was created by and on:
Author: Hector Rivas Gandara <keymon@gmail.com>
Date:   Sun Mar 10 19:21:42 2013 +0000