Remove the Docs link in the layout navbar
The Docs link doesn't work (because the route in Sinatra is specified
with a trailing slash, which might work in other configurations, but
not in the way I use it with Rails). Plus, the Docs page is about
setting up GDash, not about using it, so it probably shouldn't be
exposed so immediately.