Use graphite_graph gem from rubygems, pessimistically at version 0.0.8
This version contains, among a few other things, a patch similar
to the code that was added to the acco fork fixing problems with
unencoded characters in Graphite URLs.

This commit virtually reverts an earlier commit (9ddd112c) which used
acco's fork but only in the Gemfile and not the gemspec. I think that we
wanted to actually use the forked gem, which never actually happened
because the gemspec was never changed. (Note that it also adds a version
constraint, which was absent before.)

However, this commit does also change the gemspec as well to add the
version constraint.